size_t instead of unsigned long

+++ src/01/pi.cc
@@ -9,10 +9,10 @@
  * @param rng the random number generator used
  * @return an approximation of π
  */
-double ApproximatePi(const unsigned long N, std::mt19937 &rng) {
+double ApproximatePi(const size_t N, std::mt19937 &rng) {
   static std::uniform_real_distribution<double> dist(0.0, 1.0);
-  unsigned long i,
-                n = 0;
+  size_t i,
+         n = 0;
   double x,
          y;
   for (i = 0; i < N; i++) {
